+++ src/features/sortOrder/ui/SortOrder.tsx
@@ -1,0 +1,19 @@
+import {Select, SelectContent, SelectItem, SelectTrigger, SelectValue} from "@/shared/ui";
+
+interface Props {
+    sortOrder: string;
+    setSortOrder: (value: string) => void;
+}
+
+export const SortOrder = ({sortOrder, setSortOrder}: Props) => {
+
+    return   <Select value={sortOrder} onValueChange={setSortOrder}>
+        <SelectTrigger className="w-[180px]">
+            <SelectValue placeholder="정렬 순서" />
+        </SelectTrigger>
+        <SelectContent>
+            <SelectItem value="asc">오름차순</SelectItem>
+            <SelectItem value="desc">내림차순</SelectItem>
+        </SelectContent>
+    </Select>
+}
